requirements.txt
pandas==0.24.2
numpy==1.14.5
tensorflow-gpu==1.11.0
matplotlib==3.0.3
